refactor: move all rtk files to inside sql runner feature folder
